Extract site settings config grids
