+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.69",
-  "tag": "v3.4.69",
+  "version": "3.4.70",
+  "tag": "v3.4.70",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.68 <3.4.69"
+      ">=3.4.69 <3.4.70"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.68 first."
+    "message": "Update to v3.4.69 first."
   }
 }
